Fix, refacto and improve the interaction between the inputs and nodes
