+++ src/poc/nodeEditor/CommandNode.tsx
@@ -5,12 +5,12 @@ import { EventDialogsRef } from './EventEditorOverlay';
 import { EventIcon, IconsFromCommand } from '@components/event/EventIcon';
 import { Handle, Position } from '@xyflow/react';
 import { useHandleConnectionState } from './useHandleConnectionState';
+import { CustomHandleContainer } from './CustomHandleContainer';
 import PlusIcon from '@assets/icons/global/plus-icon.svg';
 import InfoIcon from '@assets/icons/notification/info.svg';
 import NoteIcon from '@assets/icons/global/note.svg';
 import React, { ReactNode } from 'react';
 import styled from 'styled-components';
-import { Input, MultiLineInput } from '@components/inputs';
 
 const CommandNodeContainer = styled.div`
   display: flex;
@@ -96,6 +96,10 @@ const CommandNodeContainer = styled.div`
         gap: 2px;
         border-radius: 8px;
 
+        &:hover {
+          cursor: pointer;
+        }
+
         .icon {
           color: #6c707b;
         }
@@ -124,161 +128,6 @@ const CommandNodeContainer = styled.div`
       0px 12px 16px -6px rgba(0, 0, 0, 0.04),
       0px 0px 0px 1px rgba(202, 211, 241, 0.13);
     border-radius: 13px;
-
-    ${Input},
-    ${MultiLineInput} {
-      background-color: rgba(255, 255, 255, 0.0001);
-      /* Shadow/XS-BorderSubtle */
-      box-shadow:
-        0px 3px 1px -2px rgba(38, 47, 56, 0.06),
-        0px 2px 3px rgba(38, 47, 56, 0.05),
-        0px 0px 0px 1px rgba(202, 211, 241, 0.13);
-      border-radius: 8px;
-      height: 30px;
-    }
-
-    ${MultiLineInput} {
-      height: 76px;
-    }
-  }
-`;
-
-type CustomHandleContainerProps = {
-  position: 'left' | 'right';
-};
-
-const CustomHandleContainer = styled.div<CustomHandleContainerProps>`
-  position: absolute;
-  width: 8px;
-  height: 8px;
-  top: 16px;
-  left: ${({ position }) => (position === 'left' ? '-9px' : '321px')};
-
-  .icon {
-    position: relative;
-    display: none;
-    color: ${({ theme }) => theme.colors.text400};
-    pointer-events: none;
-
-    svg {
-      width: 10px;
-      height: 10px;
-    }
-  }
-
-  .point {
-    position: relative;
-    display: none;
-    background-color: ${({ theme }) => theme.colors.text400};
-    pointer-events: none;
-    border-radius: 100%;
-    width: 8px;
-    height: 8px;
-  }
-
-  .react-flow__handle {
-    background: #202225;
-    border: 1px solid #383a40;
-    box-shadow: 0px 0px 0px 2px #181819;
-    border-radius: 100%;
-    cursor: pointer;
-  }
-
-  &[data-connected='true'] {
-    .react-flow__handle {
-      display: flex;
-      background: none;
-      width: 8px;
-      height: 8px;
-      border: 1px solid #383a40;
-      box-shadow: none;
-      align-items: center;
-      justify-content: center;
-
-      .point {
-        display: block;
-        width: 6px;
-        height: 6px;
-      }
-    }
-
-    .react-flow__handle.connectionindicator:hover {
-      display: flex;
-      box-sizing: unset;
-      background: none;
-      width: 8px;
-      height: 8px;
-      border: 2px solid #383a40;
-      box-shadow: none;
-      align-items: center;
-      justify-content: center;
-
-      .point {
-        display: block;
-        width: 6px;
-        height: 6px;
-      }
-
-      .icon {
-        display: none;
-      }
-    }
-  }
-
-  .react-flow__handle.connectionindicator:hover {
-    display: flex;
-    box-sizing: border-box;
-    width: 16px;
-    height: 16px;
-    align-items: center;
-    justify-content: center;
-
-    background: #202225;
-    border: 1px solid #383a40;
-    box-shadow: 0 -1px 0 0 #2b4c9f; // TODO: color should be dynamic
-    border-radius: 100%;
-
-    .icon {
-      display: block;
-    }
-  }
-
-  .react-flow__handle.connectingfrom,
-  .react-flow__handle.connectingto {
-    display: flex;
-    background: none;
-    width: 12px;
-    height: 12px;
-    border: 2px solid #383a40;
-    box-shadow: none;
-    align-items: center;
-    justify-content: center;
-
-    .point {
-      display: block;
-    }
-  }
-
-  .react-flow__handle.connectingfrom.connectionindicator:hover,
-  .react-flow__handle.connectingto.connectionindicator:hover {
-    background: inherit;
-    box-shadow: none;
-
-    .icon {
-      display: none;
-    }
-  }
-
-  .react-flow__handle.connectingfrom.connectionindicator:hover {
-    border: 2px solid #383a40;
-  }
-
-  .react-flow__handle.connectingto.connectionindicator {
-    border: 2px solid #2b4c9f; // TODO: color should be dynamic
-
-    .point {
-      background-color: #2b4c9f; // TODO: color should be dynamic
-    }
   }
 `;
 
@@ -286,7 +135,7 @@ type CommandNodeProps = {
   commandType: StudioEventCommandType;
   commentCount: number;
   dialogsRef?: EventDialogsRef;
-  hasError: boolean;
+  hasError?: boolean;
   nodeId: string;
   selected?: boolean;
   children: ReactNode;
@@ -335,15 +184,17 @@ export const CommandNode = ({ commandType, commentCount, dialogsRef, hasError, n
           </div>
         </div>
         <footer style={{ height: deployFooter ? '24px' : '8px' }}>
-          {hasError && (
+          {hasError ? (
             <div className="status">
               <InfoIcon className="icon" />
               <span className="label">{t('invalid_data')}</span>
             </div>
+          ) : (
+            <div />
           )}
           <div className="actions">
             {commentCount > 0 && (
-              <div className="comments">
+              <div className="comments nodrag" onClick={(e) => e.stopPropagation()} onDoubleClick={(e) => e.stopPropagation()}>
                 <NoteIcon className="icon" />
                 <span className="count">{commentCount}</span>
               </div>
